requirements.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.34.0
azure-data-tables==12.7.0
azure-identity==1.15.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.2
cvat-cli==2.3.0
cvat-sdk==2.3.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
flower==2.0.1
future==1.0.0
h11==0.16.0
httptools==0.6.4
humanize==4.12.3
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
mongoengine==0.29.1
motor==3.7.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
packaging==25.0
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.0
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
requests==2.32.3
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinydb==4.8.2
tornado==6.5.1
tqdm==4.67.1
tuspy==0.2.5
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
